Business Client Login and Register logic with Auth refactor (#14)
* Fixes

* Add business client model

* Add business client repository

* Refactors authentication logic for user types

Improves the authentication process by differentiating between user and business client logins and registrations.

This change introduces a `BaseUser` model and `BaseUserRepository` to handle both user types. It refactors the `AuthCubit` and `AuthService` to manage separate login and registration flows for users and business clients, providing more specific error handling and streamlined authentication.

The user object is now fetched during the initial authentication status check instead of relying on local storage and token decoding.

* Adds Google Maps API key support for CI

Enables the configuration of the Google Maps API key
through a new input to the prepare environment action.

This allows for dynamically setting the API key during
the build process.

* Moves AppStartup widget to its own file

Extracts the AppStartup widget into its own file for better code organization and maintainability. This change promotes modularity and improves the overall structure of the application.