Enroll and Withdraw from events (#22)
* Adds event enrollment and withdrawal functionality

Adds event enrollment and withdrawal functionality to the event details screen, allowing users to join and leave events directly from the app.

Also improves the authentication failure handling in login and register modals by resetting the authentication state to unauthenticated. This ensures a consistent user experience after failed authentication attempts.
A cache is implemented for storing Geolocation information, reducing requests to the Google Maps API and improving performance.

* Enables event enrollment/withdrawal via API

Updates the enrollment and withdrawal logic to use the API endpoints.
This change removes the local state manipulation of the event participants and relies on the API to return the updated event data.
This ensures data consistency and reflects the server-side changes accurately.